Allow user lookup by apikey
This plugin breaks the GCode Viewer as is. When using this plugin and accessing the GCode Viewer, one gets this error in the log:

TypeError: findUser() got an unexpected keyword argument 'apikey'

This commit ensures that the findUser function knows how to handle the 'apikey' parameter from ~/octoprint/server/util/__init__.py if submitted.